chore: extract lib/context.ts file
